+++ apps/web/src/views/Info/components/InfoSearch/index.tsx
@@ -21,7 +21,7 @@ import { styled } from 'styled-components'
 import { formatAmount } from 'utils/formatInfoNumbers'
 import { CurrencyLogo, DoubleCurrencyLogo } from 'views/Info/components/CurrencyLogo'
 import SaveIcon from 'views/Info/components/SaveIcon'
-import { isAddress } from 'utils'
+import { safeGetAddress } from 'utils'
 
 const Container = styled.div`
   position: relative;
@@ -324,10 +324,8 @@ const Search = () => {
                     <Flex>
                       <CurrencyLogo address={token.address} chainName={chainName} />
                       <Text ml="10px">
-                        <Text>{`${
-                          (token.address && subgraphTokenName[isAddress(token.address) || undefined]) || token.name
-                        } (${
-                          (token.address && subgraphTokenSymbol[isAddress(token.address) || undefined]) || token.symbol
+                        <Text>{`${(token.address && subgraphTokenName[safeGetAddress(token.address)]) || token.name} (${
+                          (token.address && subgraphTokenSymbol[safeGetAddress(token.address)]) || token.symbol
                         })`}</Text>
                       </Text>
                       <SaveIcon