+++ apps/web/src/views/Nft/market/components/BuySellModals/SellModal/SellStage.tsx
@@ -2,7 +2,7 @@ import { Flex, Grid, Text, Button, BinanceIcon, LinkExternal, useModal, ScanLink
 import { useTranslation } from '@pancakeswap/localization'
 import { nftsBaseUrl, pancakeBunniesAddress } from 'views/Nft/market/constants'
 import { NftToken } from 'state/nftMarket/types'
-import { getBscScanLinkForNft, isAddress } from 'utils'
+import { getBscScanLinkForNft, safeGetAddress } from 'utils'
 import EditProfileModal from 'views/Profile/components/EditProfileModal'
 import { useProfile } from 'state/profile/hooks'
 import { Divider, HorizontalDivider, RoundedImage } from '../shared/styles'
@@ -26,7 +26,9 @@ const SellStage: React.FC<React.PropsWithChildren<SellStageProps>> = ({
   const { t } = useTranslation()
   const { hasProfile } = useProfile()
   const itemPageUrlId =
-    isAddress(nftToSell.collectionAddress) === pancakeBunniesAddress ? nftToSell.attributes[0].value : nftToSell.tokenId
+    safeGetAddress(nftToSell.collectionAddress) === safeGetAddress(pancakeBunniesAddress)
+      ? nftToSell.attributes[0].value
+      : nftToSell.tokenId
 
   const [onEditProfileModal] = useModal(<EditProfileModal onSuccess={onSuccessEditProfile} />, false)